refactor: create ColumnDefinitions class and move method
